message shows up when no articles are saved or no scraped articles

+++ public/js/saved.js
@@ -79,6 +79,12 @@ var API = {
             type: "DELETE"
         });
     },
+    deleteAllNotes: function () {
+        return $.ajax({
+            url: "api/notes/",
+            type: "DELETE"
+        });
+    },
     saveNote: function (id, note) {
         return $.ajax({
             url: "api/notes/" + id,
@@ -93,24 +99,37 @@ var API = {
 var initSaved = function () {
     $("#savedResults").empty();
     API.getSaves().then(function (data) {
-        var $saves = data.map(function (artic) {
+        console.log(data);
+        if (data.length > 0) {
+            var $saves = data.map(function (artic) {
+                var $li = $("<li>");
+    
+                var deleteButton = $("<button>").addClass("buttonMargin btn-small waves-effect waves-light right deleteIt").attr("type", "submit").attr("name", "action").text("Delete From Saved");
+                var noteButton = $("<button>").addClass("buttonMargin btn-small waves-effect waves-light modal-trigger right indigo addNote").attr("type", "submit").attr("name", "action").text("Add Note").attr("data-target", "modal1");
+    
+                var title = $("<div>").text(artic.title).addClass("collapsible-header");
+    
+                var buttonDiv = $("<div>").attr("data-id", artic._id).append(deleteButton).append(noteButton)
+    
+                var span = $("<a>").attr("href", artic.link).attr("target", "_blank").append($("<span>").text(artic.summary));
+                var body = $("<div>").addClass("collapsible-body").append(span).append(buttonDiv);
+    
+                $li.append(title).append(body);
+    
+                return $li;
+            });
+            $("#savedResults").append($saves);
+        }
+
+        else {
             var $li = $("<li>");
 
-            var deleteButton = $("<button>").addClass("buttonMargin btn-small waves-effect waves-light right deleteIt").attr("type", "submit").attr("name", "action").text("Delete From Saved");
-            var noteButton = $("<button>").addClass("buttonMargin btn-small waves-effect waves-light modal-trigger right indigo addNote").attr("type", "submit").attr("name", "action").text("Add Note").attr("data-target", "modal1");
+            var title = $("<h5>").text("No Saved Articles. Go To The Home Page and Get Scraping!").addClass("center-align");
 
-            var title = $("<div>").text(artic.title).addClass("collapsible-header");
+            $li.append(title);
+            $("#savedResults").append($li);
+        }
 
-            var buttonDiv = $("<div>").attr("data-id", artic._id).append(deleteButton).append(noteButton)
-
-            var span = $("<a>").attr("href", artic.link).attr("target", "_blank").append($("<span>").text(artic.summary));
-            var body = $("<div>").addClass("collapsible-body").append(span).append(buttonDiv);
-
-            $li.append(title).append(body);
-
-            return $li;
-        });
-        $("#savedResults").append($saves);
     });
 };
 
@@ -118,26 +137,40 @@ var deleteAllSaved = function () {
     API.deleteAllSavedArt().then(function () {
         $("#savedResults").empty();
         API.getSaves().then(function (data) {
-            var $saves = data.map(function (artic) {
+            if (data.length > 0) {
+                var $saves = data.map(function (artic) {
+                    var $li = $("<li>");
+        
+                    var deleteButton = $("<button>").addClass("buttonMargin btn-small waves-effect waves-light right deleteIt").attr("type", "submit").attr("name", "action").text("Delete From Saved");
+                    var noteButton = $("<button>").addClass("buttonMargin btn-small waves-effect waves-light modal-trigger right indigo addNote").attr("type", "submit").attr("name", "action").text("Add Note").attr("data-target", "modal1");
+        
+                    var title = $("<div>").text(artic.title).addClass("collapsible-header");
+        
+                    var buttonDiv = $("<div>").attr("data-id", artic._id).append(deleteButton).append(noteButton)
+        
+                    var span = $("<a>").attr("href", artic.link).attr("target", "_blank").append($("<span>").text(artic.summary));
+                    var body = $("<div>").addClass("collapsible-body").append(span).append(buttonDiv);
+        
+                    $li.append(title).append(body);
+        
+                    return $li;
+                });
+                $("#savedResults").append($saves);
+            }
+    
+            else {
                 var $li = $("<li>");
-
-                var deleteButton = $("<button>").addClass("buttonMargin btn-small waves-effect waves-light right deleteIt").attr("type", "submit").attr("name", "action").text("Delete From Saved");
-                var noteButton = $("<button>").addClass("buttonMargin btn-small waves-effect waves-light modal-trigger right indigo addNote").attr("type", "submit").attr("name", "action").text("Add Note").attr("data-target", "modal1");
-
-                var title = $("<div>").text(artic.title).addClass("collapsible-header");
-
-                var buttonDiv = $("<div>").attr("data-id", artic._id).append(deleteButton).append(noteButton)
-
-                var span = $("<a>").attr("href", artic.link).attr("target", "_blank").append($("<span>").text(artic.summary));
-                var body = $("<div>").addClass("collapsible-body").append(span).append(buttonDiv);
-
-                $li.append(title).append(body);
-
-                return $li;
-            });
-            $("#savedResults").append($saves);
+    
+                var title = $("<h5>").text("No Saved Articles. Go To The Home Page and Get Scraping!").addClass("center-align");
+    
+                $li.append(title);
+                $("#savedResults").append($li);
+            }
         });
-    })
+    });
+    API.deleteAllNotes().then(function () {
+    });
+
 };
 
 
@@ -148,24 +181,35 @@ var deleteSavedArt = function () {
     API.deleteSaved(id).then(function (dataTwo) {
         $("#savedResults").empty();
         API.getSaves().then(function (data) {
-            var $saves = data.map(function (artic) {
+            if (data.length > 0) {
+                var $saves = data.map(function (artic) {
+                    var $li = $("<li>");
+        
+                    var deleteButton = $("<button>").addClass("buttonMargin btn-small waves-effect waves-light right deleteIt").attr("type", "submit").attr("name", "action").text("Delete From Saved");
+                    var noteButton = $("<button>").addClass("buttonMargin btn-small waves-effect waves-light modal-trigger right indigo addNote").attr("type", "submit").attr("name", "action").text("Add Note").attr("data-target", "modal1");
+        
+                    var title = $("<div>").text(artic.title).addClass("collapsible-header");
+        
+                    var buttonDiv = $("<div>").attr("data-id", artic._id).append(deleteButton).append(noteButton)
+        
+                    var span = $("<a>").attr("href", artic.link).attr("target", "_blank").append($("<span>").text(artic.summary));
+                    var body = $("<div>").addClass("collapsible-body").append(span).append(buttonDiv);
+        
+                    $li.append(title).append(body);
+        
+                    return $li;
+                });
+                $("#savedResults").append($saves);
+            }
+    
+            else {
                 var $li = $("<li>");
-
-                var deleteButton = $("<button>").addClass("buttonMargin btn-small waves-effect waves-light right deleteIt").attr("type", "submit").attr("name", "action").text("Delete From Saved");
-                var noteButton = $("<button>").addClass("buttonMargin btn-small waves-effect waves-light modal-trigger right indigo addNote").attr("type", "submit").attr("name", "action").text("Add Note").attr("data-target", "modal1");
-
-                var title = $("<div>").text(artic.title).addClass("collapsible-header");
-
-                var buttonDiv = $("<div>").attr("data-id", artic._id).append(deleteButton).append(noteButton)
-
-                var span = $("<a>").attr("href", artic.link).attr("target", "_blank").append($("<span>").text(artic.summary));
-                var body = $("<div>").addClass("collapsible-body").append(span).append(buttonDiv);
-
-                $li.append(title).append(body);
-
-                return $li;
-            });
-            $("#savedResults").append($saves);
+    
+                var title = $("<h5>").text("No Saved Articles. Go To The Home Page and Get Scraping!").addClass("center-align");
+    
+                $li.append(title);
+                $("#savedResults").append($li);
+            }
         });
     });
 };
@@ -242,6 +286,7 @@ var deleteNote = function () {
 };
 
 $("#savedResults").on("click", ".deleteIt", deleteSavedArt);
+
 $("#clearSavedArticles").on("click", deleteAllSaved);
 
 $("#savedResults").on("click", ".addNote", addNote);
